Add left margin as a prop.
