+++ src/components/home/components/Footer.tsx
@@ -1,10 +1,27 @@
-import { Grid } from "@mui/material";
+import { Grid, Typography } from "@mui/material";
 
 function Footer(): JSX.Element {
     return (
-        <Grid container sx={{color: '#fff'}}>
-            <Grid item sx={{background: 'blue'}} xs={12}>
-                asd
+        <Grid container sx={{color: '#fff', background: '#000', minHeight: '100vh'}}>
+            <Grid item xs={12}>
+                <Typography>
+                    Questions? Call 000-800-919-1694
+                </Typography>
+            </Grid>
+            <Grid xs={4}>
+
+            </Grid>
+            <Grid xs={4}>
+
+            </Grid>
+            <Grid xs={4}>
+
+            </Grid>
+            <Grid xs={12}>
+                
+            </Grid>
+            <Grid xs={12}>
+                <Typography>Netflix India</Typography>
             </Grid>
         </Grid>
     );
