+++ src/components/home/components/Header.tsx
@@ -1,15 +1,8 @@
-import { ArrowDropDown, Language } from "@mui/icons-material";
-import { AppBar, MenuItem, Select, Toolbar, SelectChangeEvent, InputAdornment, Button } from "@mui/material";
-import { useState } from "react";
+import { AppBar, Toolbar, Button } from "@mui/material";
 import { Link } from "react-router-dom";
+import LanguageOption from "./LanguageOption";
 
 function Header(): JSX.Element {
-    const [language, setLanguage] = useState<string>('English');
-
-    function changeLanguage(event: SelectChangeEvent<string>) {
-        setLanguage(event.target.value);
-    }
-
     return (
         <AppBar position='fixed' sx={{ backgroundColor: 'rgba(0, 0, 0, 0.8)', }}>
             <Toolbar>
@@ -22,17 +15,7 @@ function Header(): JSX.Element {
                         height: 40, 
                         marginLeft: '120px' }}
                 />
-                <Select
-                    startAdornment={<InputAdornment position='start' sx={{ color: '#fff' }}><Language /></InputAdornment>}
-                    endAdornment={<InputAdornment position='end' sx={{ color: '#fff' }}><ArrowDropDown /></InputAdornment>}
-                    id='language'
-                    value={language}
-                    onChange={changeLanguage}
-                    sx={{ height: 40, marginLeft: 80, color: '#fff', border: '1px solid #555' }}
-                >
-                    <MenuItem value='English'>English</MenuItem>
-                    <MenuItem value='Hindi'>Hindi</MenuItem>
-                </Select>
+                <LanguageOption />
                 <Link to='/login'>
                     <Button
                         variant="contained"
